Create Minimum XorA.java
Given two integers A and B, the task is to find an integer X such that (X XOR A) is minimum possible and the count of set bit in X is equal to the count of set bits in B.

Example 1:

Input: 
A = 3, B = 5
Output: 3
Explanation:
Binary(A) = Binary(3) = 011
Binary(B) = Binary(5) = 101
The XOR will be minimum when x = 3
i.e. (3 XOR 3) = 0 and the number of set bits in 3 is equal to the number of set bits in 5.

------------------------------------------------------------------------------------------

BASIC SOLUTION - O(N logN) Time Complexity

First calculate the count of no.of.set bits in A and B. Based on the number of set bits in A and B, we have three cases.
We know that XOR of a number with itself is zero. Considering this two points , we can get the required number.
CASE 1 : If the Number of Set bits in A and B are equal, then we can simply consider X=A because since the number of set bits in A and B are equal, then the number of set bits in X must also be equal to number of set bits in A . So If we take X=A , XOR will be minimum with the required condition of no.of.set bits.
2. CASE 2: If the Number of set bits in A is greater than number of set bits in B. In this case , we try to minimise the XOR and maintain same no.of.set bits as in B, in that case we need to eliminate the 1's contributed by MSB's in XOR.

3. CASE 3:If the Number of Set bits in A are lesser than that in B, simply set the bits of A from LSB until the number of set bits in A are equal to that of B. This is done because, we can't decrease the value of A, in order to increase set bits only thing we can do is to set the bits to right as much as possible to minimise the xor value.